playground/exports/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.5.7
cffi==1.17.1
charset-normalizer==3.1.0
colorama==0.4.6
configparser==6.0.1
Django==5.1.3
django-db-connection-pool==1.2.5
et-xmlfile==1.1.0
exrex==0.11.0
idna==3.4
lxml==5.3.0
norminette==3.3.51
numpy==1.26.4
openpyxl==3.0.10
pandas==2.2.1
pi==0.1.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pync==2.0.3
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-futures==1.0.1
soupsieve==2.5
SQLAlchemy==2.0.36
sqlparams==6.1.0
sqlparse==0.5.2
stem==1.8.2
torrequest==0.1.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.0.2
